Refactor existing code, simplifying as much as possible before changes
